When I compile this code use gcc 13.4, I notice an error "incomplete type 'detail::fnv_hash_values<long unsigned int>' used in nested name specifier", so I modify the uint32 typedef and pass test.

+++ string_hash.hpp
@@ -12,7 +12,7 @@
 
 namespace detail
 {
-	using uint32 = unsigned;
+	using uint32 = unsigned long;
 	using uint64 = unsigned long long;
 
 	template < typename H >
